chore: refactor modules chatbot
